requirements.txt
click==8.1.8
colorama==0.4.6
Flask==2.0.1
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==8.0.26
protobuf==5.29.3
Werkzeug==2.0.1
